fixed bugs in the qemu build script

+++ fetchqemu.sh
@@ -1,12 +1,20 @@
 #!/bin/bash
-rm -rf qemu
-mkdir -p qemu
-cd qemu
-wget http://wiki.qemu-project.org/download/qemu-2.0.0.tar.bz2
-tar xf qemu-2.0.0.tar.bz2
-cd qemu-2.0.0
-mv tci.c tci.c.bak
+
+if [ ! -d qemu/qemu-latest ]; then
+  rm -rf qemu
+  mkdir -p qemu
+  cd qemu
+  wget http://wiki.qemu-project.org/download/qemu-2.1.0-rc0.tar.bz2
+  tar xf qemu-2.1.0-rc0.tar.bz2
+  ln -s qemu-2.1.0-rc0 qemu-latest
+  cd qemu-latest
+  mv tci.c tci.c.bak
+  cd ../../
+fi
+
+cd qemu/qemu-latest
+rm -f tci.c
 ln -s ../../qemu_mods/tci.c tci.c
-./configure --target-list=i386-linux-user
+./configure --target-list=i386-linux-user --enable-tcg-interpreter --enable-debug-tcg --cpu=unknown
 make -j32
